% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runttestshiny.R
\name{runttestshiny}
\alias{runttestshiny}
\title{Launch ttestshiny Interface}
\usage{
runttestshiny()
}
\value{
Nothing
}
\description{
ttestshiny() loads interactive user interface built using R shiny.
}
\details{
Start ttestshiny

The interactive user interface is to provide an easy way for people who are learning ttest. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
ttestshiny()
}
}
\keyword{ttestshiny,runttestshiny}
